chore: addresing comments
