+++ application/CohortManager/src/Functions/CaasIntegration/receiveCaasFile/ProcessFileClasses/IValidateDates.cs
@@ -1,0 +1,6 @@
+using Model;
+
+public interface IValidateDates
+{
+    public bool ValidateAlleDates(Participant participant);
+}
