+++ application/CohortManager/src/Functions/Functions.sln
@@ -172,6 +172,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AddBatchToQueueTest", "..\.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RecordsProcessedTrackerTests", "..\..\..\..\tests\CaasIntegrationTests\RecordsProcessedTrackerTests\RecordsProcessedTrackerTests.csproj", "{AA6F6E1A-7511-4A65-ADCB-B86DF46223D6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ValidateDatesTests", "..\..\..\..\tests\CaasIntegrationTests\ValidateDates\ValidateDatesTests.csproj", "{FBF592EF-DB88-4ED9-A65D-0F6C4C310C87}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -426,6 +428,10 @@ Global
 		{AA6F6E1A-7511-4A65-ADCB-B86DF46223D6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AA6F6E1A-7511-4A65-ADCB-B86DF46223D6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AA6F6E1A-7511-4A65-ADCB-B86DF46223D6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FBF592EF-DB88-4ED9-A65D-0F6C4C310C87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FBF592EF-DB88-4ED9-A65D-0F6C4C310C87}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FBF592EF-DB88-4ED9-A65D-0F6C4C310C87}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FBF592EF-DB88-4ED9-A65D-0F6C4C310C87}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
